Auth mongo db

+++ MelodyMuseAPI-DotNet8/Controllers/UserController.cs
@@ -2,13 +2,17 @@
 using MelodyMuseAPI_DotNet8.Dtos;
 using MelodyMuseAPI_DotNet8.Interfaces;
 using MelodyMuseAPI_DotNet8.Services;
+using Microsoft.AspNetCore.Authentication.JwtBearer;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using System.Security.Claims;
 
 namespace MelodyMuseAPI_DotNet8.Controllers
 {
     [ApiController]
     [Route("api/u")]
+    [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
     public class UserController : Controller
     {
         private readonly IUserService _userService;
@@ -30,6 +34,12 @@ namespace MelodyMuseAPI_DotNet8.Controllers
         [HttpGet("{id}")]
         public async Task<ActionResult<User>> GetUserById(string id)
         {
+            var currentUserId = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier)?.Value;
+            if (currentUserId == null || currentUserId != id)
+            {
+                return Unauthorized("Unauthorized Access.");
+            }
+
             var user = await _userService.GetUserById(id);
             if (user == null)
             {
@@ -38,22 +48,16 @@ namespace MelodyMuseAPI_DotNet8.Controllers
             return Ok(user);
         }
 
-        // POST: api/u/register
-        [HttpPost("register")]
-        public async Task<ActionResult<User>> RegisterUser([FromBody] UserRegistrationDto registrationDto)
-        {
-            var registeredUser = await _userService.RegisterUser(registrationDto);
-            if (registeredUser == null)
-            {
-                return BadRequest("User registration failed. User may already exist.");
-            }
-            return CreatedAtAction(nameof(GetUserById), new { id = registeredUser.Id }, registeredUser);
-        }
-
         // PUT: api/u/{id}
         [HttpPut("{id}")]
         public async Task<IActionResult> UpdateUser(string id, [FromBody] UserProfileUpdateDto updateDto)
         {
+            var currentUserId = User.Claims.FirstOrDefault(c => c.Type == "UserId")?.Value;
+            if (currentUserId == null || currentUserId != id)
+            {
+                return Unauthorized("Unauthorized Access.");
+            }
+
             var result = await _userService.UpdateUser(id, updateDto);
             if (!result)
             {
@@ -66,6 +70,12 @@ namespace MelodyMuseAPI_DotNet8.Controllers
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteUser(string id)
         {
+            var currentUserId = User.Claims.FirstOrDefault(c => c.Type == "UserId")?.Value;
+            if (currentUserId == null || currentUserId != id)
+            {
+                return Unauthorized("Unauthorized Access.");
+            }
+
             var result = await _userService.DeleteUser(id);
             if (!result)
             {
